chatbot-api/requirements.txt
chromadb
fastapi
langchain-chroma
langchain-cli
langchain-community
langchain-openai
openai
unstructured
unstructured[pdf]
uvicorn
pypdf2